KTOR-7265 Remove unnecessary body copying
